add network spam test and update helpers.py with related functions
the test jumps through a lot of loops in order to make the kin python sdk
use async, since it's necessary to implement stress tests, which is what
this is test is all about.

the code is not the easiest to read as it could be,
but this will be remedied once the kin python sdk will be updated
to natively support asyncio and aiohttp python packages.

in addition, this commit add plenty of helper functions to help
with common tasks e.g. creating accounts, submitting txs, get sequences,
all in an async manner

in addition, make --passphrase and --horizon required arguments